权限接口测试

+++ auth/src/main/java/com/anan/springboot/auth/service/impl/UserServiceImpl.java
@@ -3,12 +3,15 @@ package com.anan.springboot.auth.service.impl;
 import com.anan.springboot.auth.orm.User;
 import com.anan.springboot.auth.repository.UserRepository;
 import com.anan.springboot.auth.service.UserService;
+import com.anan.springboot.core.enums.ResultEnum;
 import com.anan.springboot.core.exception.CoreException;
 import com.anan.springboot.core.orm.ResponseResult;
+import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.UnsupportedEncodingException;
 import java.util.List;
 
 /**
@@ -16,7 +19,6 @@ import java.util.List;
  * Created on 2018/8/27.
  */
 @Service
-@Transactional
 public class UserServiceImpl implements UserService {
 
   @Autowired
@@ -32,11 +34,13 @@ public class UserServiceImpl implements UserService {
     return userRepository.findById(id).get();
   }
 
+  @Transactional
   @Override
   public User save(User data) {
     return userRepository.save(data);
   }
 
+  @Transactional
   @Override
   public User update(User data) {
     return userRepository.save(data);
@@ -49,10 +53,12 @@ public class UserServiceImpl implements UserService {
     //foreach delete, if failure? jump in catch add message,	and then continue
     for (String sid : ids) {
       try{
+        if(!userRepository.existsById(Integer.parseInt(sid)))
+          throw new CoreException(ResultEnum.DELETE_SECTION);
         userRepository.deleteById(Integer.parseInt(sid));
       }catch (CoreException e){
-        result.setCode(ResponseResult.SUCCESS);
-        result.addMessage("用户名为"+userRepository.findById(Integer.parseInt(sid)).get().getAccount()+"删除失败");
+        result.setCode(e.getCode());
+        result.setMessage(e.getMessage());
       }
     }
     return result;
